requirements.txt
flask
mysql-connector-python
bcrypt
PyJWT
